SheetNavigator Added and Adding and Deleting an Exercise From a Workout is Possible
The WorkoutDetailsViewSheetNavigator allows for the WorkoutDetailsView to display various sheets depending on which button was tapped on the screen. AddExerciseView was also added to make it possible for exercises to be added to a workout. This replaced the functionality wherein the AddEditWorkoutView was used to designate exercises for a workout with a list because communicating between ViewModels when an exercise was deselected in the list was unnecessarily complex.