Added TabView to watchOS App and more
RootView and ExercisesView are now featured in the watchOS app. ExerciseDetailsView has also been added, but it doesn't do anything yet. Unnecessary code was also removed from AddExerciseView in both apps, and WorkoutDetailsView now displays the exercises for each workout in a ForEach in the watchOS app. Lastly, it is now possible to add exercises in the watchOS app.