Added Workflow Improvements to watchOS App
When adding a RepSet, the weight stepper now defaults to 60 and moves in steps of 5, and the reps count defaults to 10 instead of 5. These default values were also added to the iOS app. Also added a notification on AddEditWorkoutView for watchOS that states that a workout's date can only be altered via the iOS app. This is due to the fact that watchOS does not yet have a DatePicker.