Added Delete Confirmation Alerts on iOS and watchOS Apps
An alert is now shown to the user when they try to delete data on watchOS and iOS. To facilitate this, AlertConstants was added. This is a struct that makes it easier for the alert messages to be the same across platforms, as each property in the struct represents the message that's shown when deleting different data.