Fixed RepSet Sorting Issues
Before, RepSets were being sorted in charts by their number property. This was unreliable because this number property was not updated when sets were deleted and updated. This number property was removed and replaced with a date property, which is much more reliable. Now, all RepSet sorting is done according to this date property, and NSSortDescriptors were added to all RepSet fetch requests to automatically sort fetched RepSets by date.