watchOS State Management in WorkoutsView
Added state management in WorkoutsView to display different things depending on whether or not the user has any workouts. This involved changing the NavigationStack in WorkoutsView to a NavigationView because the .scrollDisabled(_:) modifier doesn't seem to work properly with NavigationStack after all the workouts are deleted. Additionally, RepSetsList was refactored in the watchOS app in preparation for adding state management to it as well.