Edited Bug Fix from Previous Commit
A cleaner bug fix for the bug mentioned in the previous commit was found and implemented. Now, ExercisesView and WorkoutsView both call methods to set up their FRCs in their respective .onAppear(perform:) methods. When these methods are called, the ViewStates are set such that their respective sheets do not have a missing animation upon their dismissal if they're adding the first workout or exercise in the view.