watchOS ViewState Management Bug Fix
Previously, all ViewState management was done by adding a switch statement within a Group view. This worked until a sheet was attached to that group and views were changed underneath it, causing the sheet's dismissal to not get animated when the views underneath it were altered. In a previous commit, a fix was added that took away ViewState management from FRCs to mitigate this issue, but I did not account for the fact that that would prevent realtime updates between the iOS and watchOS apps. In this commit, these Group views have been replaced with ZStacks, which fixes this issue because the sheets are now attached to ZStacks that are always present and never altered by different ViewState conditions. Sheets appear to now always dismiss with the correct animation on watchOS.