Moved ExerciseRepsList in iOS App
Reps for an exercise in a workout can now be viewed by tapping that exercise's group box within a particular workout. RepSets can now be edited and added here as well. This was done to make the design mirror the watchOS app better, and to also keep the WorkoutDetailsView from looking too crowded. Also added styling to ExerciseSetChart that makes the .foregroundStyle(.green) of a PointMark green if it's above the user's defined goal for an exercise.