Swift List Item Tapping Bug Fixed on iOS and watchOS Chart Styling
Added styling to watchOS ExerciseSetChart that turns the PointMark green if it's above the user's weight goal. Also fixed a big that was caused in the last commit by adding .buttonStyle(.plain) to List rows. This adds the correct color behavior, but it makes only the text of a List row tappable. These modifiers have been changed back to .tint(.primary) to get the desired color and tap behavior in Lists. This change was only on iOS.