Found iOS Bug and Altered ExerciseRepsView on iOS
For reasons unknown, there is a bug in ExerciseRepsView that only occurs on an iPhone 12 (physical or simulator) where the ExerciseRepsList appears scrolled down when the view is first shown. This does not occur on an iPhone 14 Pro simulator, for example, and it also doesn't occur when using a NavigationStack. However, this commit does not use a NavigationStack in WorkoutsView or ExercisesView because of the strange .navigationTitle(_:) behavior discussed in the Notes file of this commit. I will try to create a simpler example to send to Apple in a Feedback report. Also, ExerciseRepsView now has improved ViewState management that shows when Reps haven't been added to an exericse. Also, it is now possible to add a RepSet to an exercise via ExerciseRepsView's plus button in the toolbar. The view no longer dismisses itself when all RepSets are deleted, this behavior might be brought over to watchOS as well for consistency.