Modified Styling of watchOS App
On watchOS, WorkoutsView and ExercisesView no longer display a plus button in their toolbar. This was changed because that plus button was being used as a .confirmationAction, which wasn't semantically correct. Instead, a new button is shown underneath each view's .noDataFound view text. Additionally, once a list is present in either view, a Create Workout/Exercise button is shown in the default position in the toolbar, which is more semantically correct and falls more inline with the general design of watchOS buttons.