watchOS RepSetsListView Change and More Consistent Button Styling
ExerciseInWorkoutDetailsView now has a Create Set button instead of a plus button. A chevron.right SF Symbol was also added next to the chart to signify that the chart is tappable (it is now always tappable, instead of only being tappable when at least one set has been added). Additionally, it is now possible to add a RepSet via RepSetsListView. Finally, the Exercises section of WorkoutDetailsView now has a new UI that's more consistent with the rest of the app when a workout has no exercises yet.