Styled ExerciseDetailsView on iOS App
Exercise DetailsView now features styling that shows how many sets of the exericse has been performed, how many workouts it appears in, a chart that displays a version of ExerciseSetChart, and a ProgressView that shows how close the user is to their weight goal. This is a starting point, and this view will be refactored in a future commit.