Added Exercise Editing in iOS App
It is now possible to edit an exercise in the iOS App. Once this functionality was implemented, the ExerciseSetChart view that I added was not being updated in realtime in accordance with the exercise edits because it was not aware of the ExerciseDetailsViewModel, which is where the updated Exercise was being published from. As a result, adding this functionality also meant that it was necessary to add a dedicated ExerciseDetailsSetChart that could reference the ExerciseDetailsViewModel. This meant that ExerciseSetChart was eliminated entirely, and the set chart in WorkoutDetailsView is now called WorkoutDetailsExerciseSetChart because the set charts in WorkoutDetailsView also need to be aware of the exercise in WorkoutDetailsViewModel. Also added ExerciseDetailsViewHeader.