Fixed View Update Bugs on watchOS
Fixed a bug where an exercise goal wasn't updating in realtime from within WorkoutDetailsView. This was fixed by adding WorkoutDetailsExerciseSetChartViewModel, which is a dedicated view model for the set chart in WorkoutDetailsView that allows the exercise to updated as expected whenever the exercise is modified. Also fixed a bug where ExerciseDetailsGoalProgressView wasn't getting updated in realtime when its corresponding exercise was edited. This was fixed by giving it access to an ExerciseDetailsViewModel @ObservedObject property.