Added AllExerciseRepSetsView on iOS and Bug Fixes
The iOS app now includes AllExerciseRepSetsView, which is a view that displays all the RepSets for a given exercise grouped by workout. This is shown when the set chart is selected in ExerciseDetailsView. Also fixed a bug where reps from the wrong workout were being shown in WorkoutDetailsView. Finally, a bug was discovered where the sets displayed in the set chart in ExerciseDetailsView in the correct order within their workouts, but incorrectly in order of the workouts themselves. This will be fixed in the next commit.