iOS RepSet Date Saving and Updating Bug Fix
Fixed a bug where, whenever a RepSet was being created, its date property was being set to Date.now. This was not the desired behavior because, although it resulted in the correct time information for the RepSet, it did not associate the RepSet's year, month, and day info with the Workout's year, month, and day info. Additionally, createUpdatedRepSetDate(for:) is a new method in AddEditWorkoutViewModel that takes a RepSet and updates its year, month, and day information, while leaving its time information intact. This is helpful when updating a Workout's date information, because it allows for all the RepSet objects in the updated Workout's repSetArray property to stay in sync with the Workout's year, month, and day. However, the time information is kept intact so that the RepSet objects in the repSetArray property stay sorted in the correct order. The only known bug with this functionality is that the ExerciseDetailsSetChart does not update in realtime when a Workout's date is changed. This will be deferred for now, as fixing it would require a new FRC for either the Workouts associated with the exercise or for the exercise's RepSets, and such a change doesn't seem prudent at the moment.