Styled watchOS ExerciseDetailsView and Fixed Realtime Updates in AllExerciseRepSetsView
The watchOS app now features a styled ExerciseDetailsView that's consistent with its ExerciseDetailsView. It updates in realtime as well. Also, a bug was fixed where AllExerciseRepSetsView wasn't updating in real time because its controllerDidChange(_:) method was never completed. Finally, AllExerciseRepSetsView now displays a row that says 'No sets found for this workout.' when a workout is displayed but it has no sets. Previously, the viewer would simply be shown an empty Section view for that workout.