Various iOS and watchOS Bug Fixes and Improvements
This commit features an assortment of iOS and watchoS bug fixes and improvements. On watchOS, AddEditExerciseView now properly updates exercises, and its stepper moves in 5s. AddEditRepSetView properly shows the selected RepSet's weight when a RepSet is being edited. AllExerciseRepSetsView now has ViewState management on both iOS and watchOS. On iOS, horizontal padding was added to all instances of NoDataFoundView. ExerciseDetailsView now hides the 'Your Progress' GroupBox when the exercise doesn't have any RepSets associated with it. Finally, fixed a bug in ExercisesView where an error alert would be shown any time an Exercise was deleted. This was being caused by the ViewState management in the view model's FRC Delegate, which was setting the ViewState to .error anytime a nil result was returned from the controllerDidChange(_:) method.