More iOS UI Tests, Removed .interactiveDismissDisabled() Modifiers, and Added AccessibilityIdenfiers
More UI Tests have been added to the iOS app. Additionally, all sheets can now be swiped down to dismiss. This was done due to the advice in Apple's Human Interface Guidelines that says to allow swipe to dismiss for sheets. Finally, AccessibilityIdentifiers is a new Constants file that makes it easier to track .accessibilityIdentifiers associated with different UI elements.