iOS Unit Tests Started and Helper Methods for Testing Added to DataController
Basic Unit Tests have been written for the iOS App. These only include tests for the example Workout and Exercise objects, in addition to tests for the helper methods in DataController that are for testing. Also, in DataController, the generateSampleData has been modified to work properly, and a count(for:) method has been added to make it easier and more efficient to count the amount of objects in a fetch request.